Move ContestableIssuesWidget to shared
