+++ src/applications/appeals/996/components/ContestableIssuesWidget.jsx
@@ -1,36 +1,9 @@
-import React, { useEffect, useState, useRef } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
-import { Link } from 'react-router';
 
-import { VaModal } from '@department-of-veterans-affairs/component-library/dist/react-bindings';
-import set from 'platform/utilities/data/set';
-import { setData } from 'platform/forms-system/src/js/actions';
-
-import { FETCH_CONTESTABLE_ISSUES_FAILED } from '../actions';
 import { APP_NAME } from '../constants';
 
-import { IssueCard } from '../../shared/components/IssueCard';
-import {
-  ContestableIssuesLegend,
-  NoIssuesLoadedAlert,
-  NoneSelectedAlert,
-  MaxSelectionsAlert,
-  removeModalContent,
-} from '../../shared/content/contestableIssues';
-import {
-  LAST_ISSUE,
-  MAX_LENGTH,
-  REVIEW_ISSUES,
-  SELECTED,
-} from '../../shared/constants';
-import { focusIssue } from '../../shared/utils/focus';
-import { isEmptyObject } from '../../shared/utils/helpers';
-import {
-  calculateIndexOffset,
-  getSelected,
-  someSelected,
-} from '../../shared/utils/issues';
+import ContestableIssues from '../../shared/components/ContestableIssues';
 
 /**
  * ContestableIssuesWidget - Form system parameters passed into this widget
@@ -50,220 +23,11 @@ import {
  * @param {Object[]} value - array value
  * @return {JSX}
  */
-const ContestableIssuesWidget = props => {
-  const {
-    value = [],
-    id,
-    options,
-    formContext = {},
-    apiLoadStatus, // API loaded status
-    additionalIssues,
-    setFormData,
-    formData,
-  } = props;
-
-  const [showErrorModal, setShowErrorModal] = useState(false);
-  const [showRemoveModal, setShowRemoveModal] = useState(false);
-  const [removeIndex, setRemoveIndex] = useState(null);
-  const [editState] = useState(window.sessionStorage.getItem(LAST_ISSUE));
-  const hasChecked = useRef(false);
-
-  useEffect(
-    () => {
-      if (
-        !hasChecked.current &&
-        apiLoadStatus === FETCH_CONTESTABLE_ISSUES_FAILED
-      ) {
-        hasChecked.current = true;
-      }
-    },
-    [apiLoadStatus, hasChecked],
-  );
-
-  useEffect(
-    () => {
-      if (editState) {
-        focusIssue();
-      }
-    },
-    [editState],
-  );
-
-  const onReviewPage = formContext?.onReviewPage || false;
-  window.sessionStorage.setItem(REVIEW_ISSUES, onReviewPage);
-
-  // inReviewMode = true (review page view, not in edit mode)
-  // inReviewMode = false (in edit mode)
-  const inReviewMode = (onReviewPage && formContext.reviewMode) || false;
-  const showCheckbox = !onReviewPage || (onReviewPage && !inReviewMode);
-  const { submitted } = formContext;
-
-  // combine all issues for viewing
-  const items = value
-    .map(item => ({
-      ...item.attributes,
-      [SELECTED]: item[SELECTED],
-    }))
-    .concat((additionalIssues || []).filter(Boolean));
-
-  const hasSelected = someSelected(items);
-
-  if (onReviewPage && inReviewMode && items.length && !hasSelected) {
-    return (
-      <NoneSelectedAlert
-        count={items.length}
-        headerLevel={5}
-        inReviewMode={inReviewMode}
-      />
-    );
-  }
-
-  const handlers = {
-    closeModal: () => setShowErrorModal(false),
-    onChange: (index, event) => {
-      let { checked } = event.target;
-      if (checked && getSelected(formData).length + 1 > MAX_LENGTH.SELECTIONS) {
-        setShowErrorModal(true);
-        event.preventDefault(); // prevent checking
-        checked = false;
-      } else if (index < value.length) {
-        // contestable issue check toggle
-        const changedItems = set(
-          `[${index}].${SELECTED}`,
-          checked,
-          props.value,
-        );
-        props.onChange(changedItems);
-      } else {
-        // additional issue check toggle
-        const adjustedIndex = calculateIndexOffset(index, value.length);
-        const updatedAdditionalIssues = additionalIssues.map(
-          (issue, indx) =>
-            adjustedIndex === indx ? { ...issue, [SELECTED]: checked } : issue,
-        );
-        setFormData({
-          ...formData,
-          additionalIssues: updatedAdditionalIssues,
-        });
-      }
-    },
-    onShowRemoveModal: cardIndex => {
-      const adjustedIndex = calculateIndexOffset(cardIndex, value.length);
-      setRemoveIndex(adjustedIndex);
-      setShowRemoveModal(true);
-    },
-    onRemoveModalClose: () => {
-      focusIssue(null, null, `${value.length + removeIndex},remove-cancel`);
-      setShowRemoveModal(false);
-      setRemoveIndex(null);
-    },
-    onRemoveIssue: () => {
-      const updatedAdditionalIssues = additionalIssues.filter(
-        (issue, indx) => removeIndex !== indx,
-      );
-      setShowRemoveModal(false);
-      setRemoveIndex(null);
-      // setTimeout needed to allow rerender
-      setTimeout(() => {
-        focusIssue(null, null, -1);
-      });
-
-      setFormData({
-        ...formData,
-        additionalIssues: updatedAdditionalIssues,
-      });
-    },
-  };
-
-  const content = items.map((item, index) => {
-    const itemIsSelected = !!item[SELECTED];
-    const hideCard = (inReviewMode && !itemIsSelected) || isEmptyObject(item);
-
-    const cardProps = {
-      id,
-      index,
-      item,
-      key: index,
-      options,
-      showCheckbox,
-      onReviewPage,
-      onChange: handlers.onChange,
-      onRemove: handlers.onShowRemoveModal,
-    };
-
-    // Don't show un-selected ratings in review mode
-    return hideCard ? null : <IssueCard {...cardProps} />;
-  });
-
-  const showNoIssues =
-    !onReviewPage && apiLoadStatus === FETCH_CONTESTABLE_ISSUES_FAILED;
-
-  return (
-    <>
-      <div name="eligibleScrollElement" />
-      {showNoIssues && <NoIssuesLoadedAlert />}
-      {!showNoIssues &&
-        !hasSelected &&
-        (onReviewPage || submitted) && (
-          <NoneSelectedAlert
-            count={value.length}
-            headerLevel={onReviewPage ? 4 : 3}
-            inReviewMode={inReviewMode}
-          />
-        )}
-      <fieldset className="review-fieldset">
-        <ContestableIssuesLegend
-          onReviewPage={onReviewPage}
-          inReviewMode={inReviewMode}
-        />
-        <VaModal
-          clickToClose
-          status="warning"
-          modalTitle={removeModalContent.title}
-          primaryButtonText={removeModalContent.yesButton}
-          secondaryButtonText={removeModalContent.noButton}
-          onCloseEvent={handlers.onRemoveModalClose}
-          onPrimaryButtonClick={handlers.onRemoveIssue}
-          onSecondaryButtonClick={handlers.onRemoveModalClose}
-          visible={showRemoveModal}
-        >
-          <p>
-            {removeIndex !== null
-              ? removeModalContent.description(
-                  additionalIssues[removeIndex].issue,
-                )
-              : null}
-          </p>
-        </VaModal>
-        <ul className="issues vads-u-border-top--1px vads-u-border-color--gray-light">
-          {content}
-        </ul>
-        {onReviewPage && inReviewMode ? null : (
-          <Link
-            className="add-new-issue vads-c-action-link--green"
-            to={{
-              pathname: '/add-issue',
-              search: `?index=${items.length}`,
-            }}
-          >
-            Add a new issue
-          </Link>
-        )}
-        {showErrorModal && (
-          <MaxSelectionsAlert
-            showModal
-            closeModal={handlers.closeModal}
-            appName={APP_NAME}
-          />
-        )}
-      </fieldset>
-    </>
-  );
-};
+const ContestableIssuesWidget = props => (
+  <ContestableIssues {...props} appName={APP_NAME} />
+);
 
 ContestableIssuesWidget.propTypes = {
-  additionalIssues: PropTypes.array,
-  apiLoadStatus: PropTypes.string,
   formContext: PropTypes.shape({
     onReviewPage: PropTypes.bool,
     reviewMode: PropTypes.bool,
@@ -271,26 +35,12 @@ ContestableIssuesWidget.propTypes = {
   }),
   formData: PropTypes.shape({
     contestedIssues: PropTypes.array,
+    additionalIssues: PropTypes.array,
   }),
   id: PropTypes.string,
   options: PropTypes.shape({}),
-  setFormData: PropTypes.func,
-  value: PropTypes.array,
-  onChange: PropTypes.func,
+  value: PropTypes.array, // not used
+  onChange: PropTypes.func, // not used
 };
 
-const mapStateToProps = state => ({
-  formData: state.form?.data || {},
-  apiLoadStatus: state.contestableIssues?.status || '',
-  additionalIssues: state.form?.data.additionalIssues || [],
-});
-const mapDispatchToProps = {
-  setFormData: setData,
-};
-
-export { ContestableIssuesWidget };
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps,
-)(ContestableIssuesWidget);
+export default ContestableIssuesWidget;
