+++ src/applications/appeals/shared/actions/index.js
@@ -1,0 +1,5 @@
+export const FETCH_CONTESTABLE_ISSUES_INIT = 'FETCH_CONTESTABLE_ISSUES_INIT';
+export const FETCH_CONTESTABLE_ISSUES_SUCCEEDED =
+  'FETCH_CONTESTABLE_ISSUES_SUCCEEDED';
+export const FETCH_CONTESTABLE_ISSUES_FAILED =
+  'FETCH_CONTESTABLE_ISSUES_FAILED';
